// src/redux/auth/types.ts
export enum AuthActionTypes {
  SET_LOGGED_IN_USER = 'SET_LOGGED_IN_USER',
  SET_LOGGED_OUT_USER = 'SET_LOGGED_OUT_USER',
  SET_IS_SIGNING = 'SET_IS_SIGNING',
  RESET_IS_SIGNING = 'RESET_IS_SIGNING',
  SET_FORGET_PASSWORD_USER_EMAIL = 'SET_FORGET_PASSWORD_USER_EMAIL',
  RESET_FORGET_PASSWORD_USER_EMAIL = 'RESET_FORGET_PASSWORD_USER_EMAIL',
  REFRESH_STATE = 'REFRESH_STATE',
}